+++ cmark-extra/extensions/extra-extensions.c
@@ -17,7 +17,10 @@
 #include "inlineimage.h"
 #include "emoji.h"
 #include "heads.h"
+#include "highlight.h"
 #include "math_ext.h"
+#include "sub_ext.h"
+#include "sup_ext.h"
 
 static int extra_extensions_registration(cmark_plugin *plugin) {
     cmark_plugin_register_syntax_extension(plugin, create_mention_extension());
@@ -28,7 +31,10 @@ static int extra_extensions_registration(cmark_plugin *plugin) {
 
     cmark_plugin_register_syntax_extension(plugin, create_emoji_extension());
     cmark_plugin_register_syntax_extension(plugin, create_heads_extension());
+    cmark_plugin_register_syntax_extension(plugin, create_highlight_extension());
     cmark_plugin_register_syntax_extension(plugin, create_math_extension());
+    cmark_plugin_register_syntax_extension(plugin, create_sup_extension());
+    cmark_plugin_register_syntax_extension(plugin, create_sub_extension());
     return 1;
 }