+++ cmark-extra/extensions/sub_ext.h
@@ -1,0 +1,16 @@
+//
+//  sub_ext.h
+//  QLMarkdown
+//
+//  Created by Sbarex on 04/08/24.
+//
+
+#ifndef sub_ext_h
+#define sub_ext_h
+
+#include <stdio.h>
+#include "cmark-gfm-core-extensions.h"
+
+cmark_syntax_extension *create_sub_extension(void);
+
+#endif /* sub_ext_h */