+++ cmark-extra/extensions/sup_ext.h
@@ -1,0 +1,16 @@
+//
+//  sup_ext.h
+//  QLMarkdown
+//
+//  Created by Sbarex on 04/08/24.
+//
+
+#ifndef sup_ext_h
+#define sup_ext_h
+
+#include <stdio.h>
+#include "cmark-gfm-core-extensions.h"
+
+cmark_syntax_extension *create_sup_extension(void);
+
+#endif /* sup_ext_h */